fix unsafe memory use in debug build
I focused on the Rust and thought this was correct (in #5), but I
misssed the obvious JS part--`new Uint8Array(buffer, offset, len)`
creates a _view_ to the underlying array, it doesn't copy in. In the
production build this was fine, but in the debug build the heap
at the returned location happens to be overwritten.

This fixes that without performance penalty (since we were doing this
anyway later on) by having the consumer pass in memory to copy into.